refactor(FOROME-871): move part UI from body to drawer window component
