+++ src/components/variant/ui/body.tsx
@@ -1,16 +1,11 @@
 import { Dispatch, ReactElement, SetStateAction, useEffect } from 'react'
 import GridLayout from 'react-grid-layout'
-import Checkbox from 'react-three-state-checkbox'
-import { clone, get } from 'lodash'
 import { observer } from 'mobx-react-lite'
 
 import { IGridLayout } from '@declarations'
 import { SessionStoreManager } from '@core/storage-management/session-store-manager'
-import { t } from '@i18n'
 import variantStore from '@store/variant'
-import { Icon } from '@ui/icon'
 import { DrawerWindow } from './drawer-window'
-import { IgvButton } from './igv-button'
 
 interface Props {
   drawerWidth: number
@@ -80,90 +75,11 @@ export const VariantBody = observer(
               id="parent"
               className="flex flex-col grid-item-dark flex-shrink-0 bg-blue-dark rounded text-grey-blue mb-2 text-14 leading-16px overflow-hidden pb-3"
             >
-              <div
-                className="flex justify-between p-3 rounded-t font-bold text-white uppercase cursor-pointer hover:bg-blue-bright"
-                onClick={e => {
-                  const target = e.target as HTMLButtonElement
-
-                  if (
-                    target &&
-                    target.classList.contains('dragHandleSelector')
-                  ) {
-                    return
-                  }
-
-                  const cloneRecords: any = variantStore.recordsDisplayConfig
-
-                  const drawerElement = document.querySelector(
-                    `#drawer-${aspect.name}`,
-                  )
-
-                  const clientHeight = get(
-                    drawerElement?.firstChild,
-                    'clientHeight',
-                    0,
-                  )
-
-                  const openedH = clientHeight * 0.0208 + 1.3
-
-                  setLayout((prev: IGridLayout[]) => {
-                    const clonedLayout: any[] = clone(prev)
-
-                    const layoutItemIndex = clonedLayout.findIndex(
-                      (aspectItem: any) => aspectItem.i === aspect.name,
-                    )
-
-                    clonedLayout[layoutItemIndex].h = cloneRecords[aspect.name]
-                      .isOpen
-                      ? 1
-                      : openedH
-
-                    variantStore.updateRecordsDisplayConfig(
-                      clonedLayout[layoutItemIndex].i,
-                      clonedLayout[layoutItemIndex].h,
-                    )
-
-                    const reflowLayout = clonedLayout.map(
-                      (layoutItem, layoutIndex: number) => {
-                        if (layoutIndex < layoutItemIndex) {
-                          return layoutItem
-                        }
-
-                        return {
-                          ...layoutItem,
-                          y: layoutItem.y + openedH,
-                        }
-                      },
-                    )
-
-                    window.sessionStorage.setItem(
-                      'gridLayout',
-                      JSON.stringify(reflowLayout),
-                    )
-
-                    return reflowLayout
-                  })
-                  variantStore.checkRecodsDisplaying()
-                }}
-              >
-                {aspect.title}
-                <div className="flex">
-                  {aspect.name === 'view_gen' && <IgvButton />}
-                  {aspect.name === 'view_transcripts' && (
-                    <div>
-                      <span>{t('variant.showSelectionOnly')}</span>
-                      <Checkbox checked={false} />
-                    </div>
-                  )}
-
-                  <Icon
-                    name="ArrowsOut"
-                    className="dragHandleSelector mr-1 cursor-move hover:text-blue-bright"
-                  />
-                </div>
-              </div>
-
-              <DrawerWindow aspect={aspect} layout={layout} />
+              <DrawerWindow
+                aspect={aspect}
+                layout={layout}
+                setLayout={setLayout}
+              />
             </div>
           )
         })}
